requirements.txt
streamlit==1.38.0
pandas==2.2.2
neo4j==5.24.0
requests==2.32.3
scikit-learn==1.5.1
tensorflow==2.17.0
nltk==3.9.1
pillow==10.4.0
python-dotenv==1.0.1

